refactor(Store): 도메인 내부에 internal 폴더 생성 후, 파일 경로 리팩토링
