+++ src/main/java/app/domain/store/internal/InternalStoreController.java
@@ -1,10 +1,11 @@
-package app.domain.store;
+package app.domain.store.internal;
 
 import java.util.UUID;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -34,4 +35,6 @@ public class InternalStoreController {
 		@PathVariable Long userId) {
 		return internalStoreService.isStoreOwner(storeId, userId);
 	}
+
+
 }
